requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.1

asyncpg>=0.29.0
redis>=5.0.1

dask>=2024.1.0
distributed>=2024.1.0
psutil>=5.9.7

opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation>=0.42b0
pyinstrument>=4.6.1

onnxruntime>=1.16.3
numpy>=1.26.3

Pillow>=10.2.0
opencv-python>=4.9.0.80

playwright>=1.41.0
aiofiles>=23.2.1

jishaku>=2.5.1
typing_extensions>=4.9.0
colorama>=0.4.6
boltons>=23.1.1
python-dateutil>=2.8.2
humanize>=4.9.0

black>=24.1.1
isort>=5.13.2
mypy>=1.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.5